Fixes to server, test, config.yml Ran shrinkwrap Edited travis.yml to install build-essential 	modified:   .gitignore 	modified:   .travis.yml 	modified:   bin/server.js 	modified:   config.yml 	new file:   npm-shrinkwrap.json 	renamed:    test/events.js -> test/event_test.js 	modified:   test/runtests.sh
