+++ .gitignore
@@ -1,2 +1,3 @@
 /taskcluster-events.conf.json
 /node_modules/
+/user-config.yml
